refactoring base class of metalearners to generalize finetuning and eval procedures in base class
