+++ src/metalearners/maml.py
@@ -20,14 +20,9 @@ from collections import defaultdict
 from .base import BaseMetaLearner
 from ..taskheads import TaskHead
 from ..utils import move_to_device
-from ..datasets import NLUDataLoader
 
 # typing imports 
-from typing import Dict, Tuple, List, Callable, Union
-from ..datasets import NLUDataset
-from ..utils.data import ShuffledIterableDataset
-from ..evaluation.evaluator import Metric
-
+from typing import Dict, List
 
 logger = logging.getLogger(__name__)
 
@@ -317,229 +312,3 @@ class MAML(BaseMetaLearner):
         )
 
         return loss
-
-
-    def run_evaluation(
-        self, 
-        finetune_dataset: ShuffledIterableDataset,
-        eval_dataset: NLUDataset,
-        metric: Metric,
-        task_type: str, 
-        task_head_init_method: str,
-        num_classes: int,
-        max_epochs: int = 10,
-        batch_size: int = 256,
-        device: torch.device = None,
-        return_finetune_info: bool = True,
-    ) -> Union[Tuple[float, float], Tuple[float, float, Dict[str, List[float]]]]:
-        """
-        Runs finetuning of the model on the support set data stored in the support_dataloader,
-        and then evaluates the model on the eval_dataloader.
-
-        Args: 
-            * finetune_dataset: An NLU Dataset containing the data for finetuning the
-                pre-trained model on a given NLU task
-            * eval_dataset: An NLU Dataset containing the evaluation set data for the given NLU 
-                task
-            * metric: A callable metric instance that uses the finetuned model to compute the
-                target evaluation metric on the eval_dataloader
-            * task_type: A string indicating the type of task (e.g. 'classification', 'qa', etc.)
-            * task_head_init_method: A string indicating the method used to initialize the task
-                head weights
-            * num_classes: An integer indicating the number of classes in the task
-            * max_epochs: An integer indicating the max number of epochs to run the finetuning;
-                we break out of the loop if the validation loss stops improving
-            * batch_size: An integer indicating the batch size to use for the finetuning
-            * device: Optional string to specify a device to override base_device
-            * return_finetune_info: A boolean indicating whether to return the training losses
-                and accuracies for the finetuning process
-
-        Returns (if return_finetune_info is False):
-            * eval_metric: The evaluation metric for the given task
-            * total_eval_loss: A float indicating the loss of the model on the evaluation set 
-        Returns (if return_finetune_info is True):
-            * eval_metric: The evaluation metric for the given 
-            * total_eval_loss: A float indicating the loss of the model on the evaluation set
-            * finetuning_info: A dictionary containing the following information:
-                * train_losses: A list of training losses for every evaluation step
-                * train_accuracies: A list of training accuracies for every evaluation step
-                * val_losses: A list of validation losses for every evaluation step
-                * val_accuracies: A list of validation accuracies for every evaluation step
-        """
-
-        assert 'protomaml' not in task_head_init_method,\
-            "Protomaml task head initialization is not supported for evaluation"
-        
-        if device is None:
-            device = self.base_device
-
-        self.base_model.to(device)
-
-        # Setting up the task head for the task
-        with torch.no_grad(): 
-            init_kwargs = self.get_task_init_kwargs(
-                task_type,
-                task_head_init_method,
-                num_classes,
-                device=device
-            )
-            task_head_weights = TaskHead.initialize_task_head(**init_kwargs)
-
-        finetune_model = copy.deepcopy(self.base_model)
-
-        # Training parameters - should not be touched by the user
-        MAX_PATIENCE = 3
-        EVAL_EVERY_N_STEPS = 20
-        INITIAL_LR = 1e-5
-
-        patience = MAX_PATIENCE
-        best_dev_metric = None
-
-        dev_batch = move_to_device(finetune_dataset.dev_batch, device)
-
-        # Setting up the optimizer
-        finetune_optimizer_param_groups = self.finetune_optimizer_param_groups(
-            finetune_model,
-            task_head_weights,
-            add_decay_information=True,
-            weight_decay_val=0.0
-        )
-
-        finetune_optimizer = AdamW(finetune_optimizer_param_groups, lr=INITIAL_LR)
-
-        finetune_model.train()
-
-        total_step_num = 0
-
-        early_exit_training = False
-
-        if return_finetune_info:
-            # Setting up the training info dictionary
-            finetune_info = [] 
-
-        for epoch in range(max_epochs):
-
-            if early_exit_training:
-                break
-            
-            finetune_dataloader = NLUDataLoader(
-                finetune_dataset,
-                batch_size=batch_size
-            )
-
-            # Finetune the model on the data in the finetune dataloader 
-            for finetune_batch in finetune_dataloader:
-
-                finetune_optimizer.zero_grad()
-
-                finetune_batch = move_to_device(finetune_batch, device)
-
-                outputs = finetune_model(
-                    input_ids=finetune_batch['input_ids'],
-                    attention_mask=finetune_batch['attention_mask']
-                )
-
-                _, loss = self._compute_task_loss(
-                    outputs,
-                    finetune_batch,
-                    task_head_weights, 
-                    task_type=task_type
-                )
-
-                if total_step_num % EVAL_EVERY_N_STEPS == 0:
-                    # Evaluating the model on the dev set to possbily break out early
-                    with torch.no_grad():
-                        finetune_model.eval()
-
-                        outputs = finetune_model(
-                            input_ids=dev_batch['input_ids'],
-                            attention_mask=dev_batch['attention_mask']
-                        )
-
-                        dev_logits, dev_loss = self._compute_task_loss(
-                            outputs,
-                            dev_batch,
-                            task_head_weights,
-                            task_type=task_type
-                        )
-                        
-                        dev_predictions = torch.argmax(dev_logits, dim=-1).tolist()
-                        dev_labels = dev_batch['label_ids'].tolist()
-
-                        dev_metric = metric(dev_predictions, dev_labels)
-
-                        if best_dev_metric is None or \
-                            metric.summary(dev_metric, best_dev_metric) == dev_metric:
-
-                            best_dev_metric = dev_metric
-                            patience = MAX_PATIENCE
-                        else:
-                            patience -= 1
-                            if patience == 0:
-                                early_exit_training = True 
-
-                        finetune_model.train()
-
-                        if return_finetune_info:
-                            finetune_info.append({
-                                'train_loss': loss.item(),
-                                'dev_loss': dev_loss.item(),
-                                'dev_metric': dev_metric,
-                                'step_num': total_step_num
-                            })
-
-                if early_exit_training:
-                    break
-
-                loss.backward()
-                finetune_optimizer.step()
-
-                total_step_num += 1
-
-
-        # Running full evaluation
-        finetune_model.eval()
-
-        eval_labels = []
-        eval_predictions = []
-    
-        total_eval_loss = 0.0
-        total_eval_samples = 0
-
-        eval_dataloader = NLUDataLoader(
-            eval_dataset,
-            batch_size=batch_size
-        )
-
-        with torch.no_grad():
-
-            for eval_batch in eval_dataloader: 
-                eval_batch = move_to_device(eval_batch, device)
-
-                eval_outputs = finetune_model(
-                    input_ids=eval_batch['input_ids'],
-                    attention_mask=eval_batch['attention_mask'],
-                )
-
-                eval_logits, eval_loss = self._compute_task_loss(
-                    eval_outputs, 
-                    eval_batch,
-                    task_head_weights,
-                    task_type=task_type
-                )
-
-                eval_predictions.extend(torch.argmax(eval_logits, dim=-1).tolist())
-                eval_labels.extend(eval_batch['label_ids'].tolist())
-
-                batch_size = eval_logits.size(0)
-                total_eval_loss += eval_loss.detach().item() * batch_size # loss avg across batch
-                total_eval_samples += batch_size 
-
-            total_eval_loss /= total_eval_samples
-
-            eval_metric = metric(eval_predictions, eval_labels)
-
-        if return_finetune_info:
-            return (eval_metric, total_eval_loss, finetune_info)
-        else:
-            return (eval_metric, total_eval_loss)
